Revert 191364 "cc: Add ‘chromium_code’: 1 to cc.gyp and cc_t..."
> cc: Add ‘chromium_code’: 1 to cc.gyp and cc_tests.gyp
> 
> And fix compile errors that it causes.
> 
> Review URL: https://codereview.chromium.org/13206004

TBR=danakj@chromium.org
Review URL: https://codereview.chromium.org/13316003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@191371 0039d316-1c4b-4281-b951-d872f2087c98

+++ cc/layers/scrollbar_layer_unittest.cc
@@ -207,7 +207,7 @@ TEST(ScrollbarLayerTest, SolidColorDrawQuads) {
     scrollbar_layer_impl->AppendQuads(&quad_culler, &data);
 
     const QuadList& quads = quad_culler.quad_list();
-    ASSERT_EQ(1u, quads.size());
+    ASSERT_EQ(1, quads.size());
     EXPECT_EQ(DrawQuad::SOLID_COLOR, quads[0]->material);
     EXPECT_RECT_EQ(gfx::Rect(1, 0, 4, 3), quads[0]->rect);
   }
@@ -221,7 +221,7 @@ TEST(ScrollbarLayerTest, SolidColorDrawQuads) {
     scrollbar_layer_impl->AppendQuads(&quad_culler, &data);
 
     const QuadList& quads = quad_culler.quad_list();
-    ASSERT_EQ(1u, quads.size());
+    ASSERT_EQ(1, quads.size());
     EXPECT_EQ(DrawQuad::SOLID_COLOR, quads[0]->material);
     EXPECT_RECT_EQ(gfx::Rect(2, 0, 8, 6), quads[0]->rect);
   }
@@ -239,7 +239,7 @@ TEST(ScrollbarLayerTest, SolidColorDrawQuads) {
     scrollbar_layer_impl->AppendQuads(&quad_culler, &data);
 
     const QuadList& quads = quad_culler.quad_list();
-    ASSERT_EQ(1u, quads.size());
+    ASSERT_EQ(1, quads.size());
     EXPECT_EQ(DrawQuad::SOLID_COLOR, quads[0]->material);
     EXPECT_RECT_EQ(gfx::Rect(4, 0, 2, 3), quads[0]->rect);
   }
@@ -271,7 +271,7 @@ TEST(ScrollbarLayerTest, LayerDrivenSolidColorDrawQuads) {
     scrollbar_layer_impl->AppendQuads(&quad_culler, &data);
 
     const QuadList& quads = quad_culler.quad_list();
-    ASSERT_EQ(1u, quads.size());
+    ASSERT_EQ(1, quads.size());
     EXPECT_EQ(DrawQuad::SOLID_COLOR, quads[0]->material);
     EXPECT_RECT_EQ(gfx::Rect(4, 0, 2, 3), quads[0]->rect);
   }
@@ -351,7 +351,7 @@ class ScrollbarLayerTestResourceCreation : public testing::Test {
   ScrollbarLayerTestResourceCreation()
       : fake_client_(FakeLayerTreeHostClient::DIRECT_3D) {}
 
-  void TestResourceUpload(size_t expected_resources) {
+  void TestResourceUpload(int expected_resources) {
     layer_tree_host_.reset(
         new MockLayerTreeHost(&fake_client_, layer_tree_settings_));
 
@@ -394,7 +394,7 @@ class ScrollbarLayerTestResourceCreation : public testing::Test {
     scrollbar_layer->SetTexturePriorities(calculator);
     layer_tree_host_->contents_texture_manager()->PrioritizeTextures();
     scrollbar_layer->Update(&queue, &occlusion_tracker, NULL);
-    EXPECT_EQ(0u, queue.FullUploadSize());
+    EXPECT_EQ(0, queue.FullUploadSize());
     EXPECT_EQ(expected_resources, queue.PartialUploadSize());
 
     testing::Mock::VerifyAndClearExpectations(layer_tree_host_.get());
